\name{gcm-package}
\alias{gcm-package}
\alias{gcm}
\docType{package}
\title{
Use GRASS colormaps in raster plots
}
\description{
Use GRASS-type colormap definitions as raster map colormaps
}
\details{
\tabular{ll}{
Package: \tab gcm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-21\cr
License: \tab GPL (>= 3)\cr
}
provides a plot and a legend method sitting on top of raster::plot using 
GRASS-style colormaps allowing for discontinous color gradients 
(see \href{http://grass.osgeo.org/grass64/manuals/r.colors.html}{r.colors})

}
\author{
Jan M. Holstein <holstein@uni-bremen.de>

Maintainer: Jan M. Holstein <holstein@uni-bremen.de>}
\references{
\link[raster]{plot}
}
\keyword{spatial}
\seealso{
\code{\link[gcm]{plot.gcm}, \link[gcm]{plot.gcm.legend}}
}

